5.3.3: IDP committing to rel ePSA on request suffices
The registrar only needs to check the IDP's committment, not the factual release in specific cases. So this seems fine.

Also, ePSA is an essential part of the value proposition of academic identity federations. There's nothing wrong with academic institutions releasing ePSA (which itself is not personal data) to those Service Providers that request it, e.g. via RequestedAttribute elements or Entity Attributes, such as REFEDS R&S.